Trigger links check only when a new node is inserted
TeuxDeux doesn't allow updating items and dragging items to the input + re-adding is the only way to modify items.

Not-triggering the check on tree modify fixes a bug, when an item is longer and the ellipsis are shown. On ellipsis hover the whole wrap is hidden, the link add code is triggered and it messes up the margins.